+++ Group2WPF/BookingWPF.xaml.cs
@@ -107,15 +107,20 @@ namespace Group2WPF
         {
             try
             {
-                if (txtId.Text.Length > 0)
-                {
-                    Booking booking = new Booking();
-                    booking.Id = Int32.Parse(txtId.Text);
-                    booking.PassengerId = Int32.Parse(cboPassenger.SelectedValue.ToString());
-                    booking.FlightId = Int32.Parse(cboFlight.SelectedValue.ToString());
-                    booking.BookingPlatformId = Int32.Parse(cboBookingPlatform.SelectedValue.ToString());
-                    booking.BookingTime = DateTime.Parse(txtBookingTime.Text);
-                    _bookingService.updateBooking(booking);
+                var result = MessageBox.Show("Are you sure you want to Update this Booking?", "Confirm update", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes) {
+                    if (txtId.Text.Length > 0)
+                    {
+                        Booking booking = new Booking();
+                        booking.Id = Int32.Parse(txtId.Text);
+                        booking.PassengerId = Int32.Parse(cboPassenger.SelectedValue.ToString());
+                        booking.FlightId = Int32.Parse(cboFlight.SelectedValue.ToString());
+                        booking.BookingPlatformId = Int32.Parse(cboBookingPlatform.SelectedValue.ToString());
+                        booking.BookingTime = DateTime.Parse(txtBookingTime.Text);
+                        _bookingService.updateBooking(booking);
+                        MessageBox.Show("Update successfully", "Success");
+                    }
+
                 }
                 else
                 {
@@ -137,19 +142,23 @@ namespace Group2WPF
         {
             try
             {
-                if (_bookingService.getBookingById(Int32.Parse(txtId.Text)) != null)
-                {
-                    MessageBox.Show("Code existed!", "ERROR");
-                }
-                else
-                {
-                    Booking booking = new Booking();
-                    booking.Id = Int32.Parse(txtId.Text);
-                    booking.PassengerId = Int32.Parse(cboPassenger.SelectedValue.ToString());
-                    booking.FlightId = Int32.Parse(cboFlight.SelectedValue.ToString());
-                    booking.BookingPlatformId = Int32.Parse(cboBookingPlatform.SelectedValue.ToString());
-                    booking.BookingTime = DateTime.Parse(txtBookingTime.Text);
-                    _bookingService.addBooking(booking);
+                var result = MessageBox.Show("Are you sure you want to create this Booking?", "Confirm create", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes) {
+                    if (_bookingService.getBookingById(Int32.Parse(txtId.Text)) != null)
+                    {
+                        MessageBox.Show("Code existed!", "ERROR");
+                    }
+                    else
+                    {
+                        Booking booking = new Booking();
+                        booking.Id = Int32.Parse(txtId.Text);
+                        booking.PassengerId = Int32.Parse(cboPassenger.SelectedValue.ToString());
+                        booking.FlightId = Int32.Parse(cboFlight.SelectedValue.ToString());
+                        booking.BookingPlatformId = Int32.Parse(cboBookingPlatform.SelectedValue.ToString());
+                        booking.BookingTime = DateTime.Parse(txtBookingTime.Text);
+                        _bookingService.addBooking(booking);
+                        MessageBox.Show("Add successfully", "Success");
+                    }
                 }
             }
             catch (Exception ex)
@@ -167,15 +176,19 @@ namespace Group2WPF
         {
             try
             {
-                if (txtId.Text.Length > 0)
-                {
-                    Booking booking = new Booking();
-                    booking.Id = Int32.Parse(txtId.Text);
-                    booking.PassengerId = Int32.Parse(cboPassenger.SelectedValue.ToString());
-                    booking.FlightId = Int32.Parse(cboFlight.SelectedValue.ToString());
-                    booking.BookingPlatformId = Int32.Parse(cboBookingPlatform.SelectedValue.ToString());
-                    booking.BookingTime = DateTime.Parse(txtBookingTime.Text);
-                    _bookingService.removeBooking(booking);
+                var result = MessageBox.Show("Are you sure you want to Delete this airline?", "Confirm delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes) {
+                    if (txtId.Text.Length > 0)
+                    {
+                        Booking booking = new Booking();
+                        booking.Id = Int32.Parse(txtId.Text);
+                        booking.PassengerId = Int32.Parse(cboPassenger.SelectedValue.ToString());
+                        booking.FlightId = Int32.Parse(cboFlight.SelectedValue.ToString());
+                        booking.BookingPlatformId = Int32.Parse(cboBookingPlatform.SelectedValue.ToString());
+                        booking.BookingTime = DateTime.Parse(txtBookingTime.Text);
+                        _bookingService.removeBooking(booking);
+                        MessageBox.Show("Delete successfully", "Success");
+                    }
                 }
                 else
                 {
