+++ Group2WPF/PassengerWPF.xaml.cs
@@ -74,42 +74,47 @@ namespace Group2WPF
 		private void Insert_Click(object sender, RoutedEventArgs e)
 		{
 			try
-			{
-				if (ValidateData())
-				{
-					int id = Int32.Parse(txtID.Text);
-					string fname = txtFName.Text;
-					string lname = txtLName.Text;
-					string gmail = txtGmail.Text;
-					string country = txtCountry.Text;
-					string gender = "";
-					DateOnly dob = DateOnly.Parse(dpDOB.Text);
-					if (rbFemale.IsChecked == true)
+            {
+                var result = MessageBox.Show("Are you sure you want to Delete this flight?", "Confirm delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+					if (ValidateData())
 					{
-						gender = "Female";
+						int id = Int32.Parse(txtID.Text);
+						string fname = txtFName.Text;
+						string lname = txtLName.Text;
+						string gmail = txtGmail.Text;
+						string country = txtCountry.Text;
+						string gender = "";
+						DateOnly dob = DateOnly.Parse(dpDOB.Text);
+						if (rbFemale.IsChecked == true)
+						{
+							gender = "Female";
+						}
+						else
+						{
+							gender = "Male";
+						}
+						Passenger? pp = _passengerService.Get(id);
+						if (pp != null)
+						{
+							MessageBox.Show("duplicate");
+							return;
+						}
+						Passenger p = new Passenger()
+						{
+							Id = id,
+							FirstName = fname,
+							LastName = lname,
+							Email = gmail,
+							Country = country,
+							Gender = gender,
+							DateOfBirth = dob,
+						};
+						_passengerService.Add(p);
+						MessageBox.Show("Add successfully", "Success");
+						LoadPassenger();
 					}
-					else
-					{
-						gender = "Male";
-					}
-					Passenger? pp = _passengerService.Get(id);
-					if(pp != null)
-					{
-						MessageBox.Show("duplicate");
-						return;
-					}
-					Passenger p = new Passenger()
-					{
-						Id = id,
-						FirstName = fname,
-						LastName = lname,
-						Email = gmail,
-						Country = country,
-						Gender = gender,
-						DateOfBirth = dob,
-					};
-					_passengerService.Add(p);
-					LoadPassenger();
 				}
 				else
 				{
@@ -127,45 +132,50 @@ namespace Group2WPF
 		{
 			try
 			{
-				if (ValidateData())
-				{
-					int id = Int32.Parse(txtID.Text);
-					Passenger? p = _passengerService.Get(id);
-					if (p == null)
+                var result = MessageBox.Show("Are you sure you want to Delete this flight?", "Confirm delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+					if (ValidateData())
 					{
-						MessageBox.Show("Not found passenger", "Fail!");
-					}
-					else
-					{
-						string fname = txtFName.Text;
-						string lname = txtLName.Text;
-						string gmail = txtGmail.Text;
-						string country = txtCountry.Text;
-						string gender = "";
-						DateOnly dob = DateOnly.Parse(dpDOB.Text);
-						if (rbFemale.IsChecked == true)
+						int id = Int32.Parse(txtID.Text);
+						Passenger? p = _passengerService.Get(id);
+						if (p == null)
 						{
-							gender = "Female";
+							MessageBox.Show("Not found passenger", "Fail!");
 						}
 						else
 						{
-							gender = "Male";
+							string fname = txtFName.Text;
+							string lname = txtLName.Text;
+							string gmail = txtGmail.Text;
+							string country = txtCountry.Text;
+							string gender = "";
+							DateOnly dob = DateOnly.Parse(dpDOB.Text);
+							if (rbFemale.IsChecked == true)
+							{
+								gender = "Female";
+							}
+							else
+							{
+								gender = "Male";
+							}
+							p = new Passenger()
+							{
+								Id = id,
+								FirstName = fname,
+								LastName = lname,
+								Email = gmail,
+								Country = country,
+								Gender = gender,
+								DateOfBirth = dob,
+							};
+							_passengerService.Update(p);
+							MessageBox.Show("Update successfully", "Success");
+							txtID.IsReadOnly = false;
+							ResetInput();
+							txtFilterFNameOrLastName.Text = string.Empty;
+							dpFilterDOB.Text = string.Empty;
 						}
-						p = new Passenger()
-						{
-							Id = id,
-							FirstName = fname,
-							LastName = lname,
-							Email = gmail,
-							Country = country,
-							Gender = gender,
-							DateOfBirth = dob,
-						};
-						_passengerService.Update(p);
-						txtID.IsReadOnly = false;
-						ResetInput();
-						txtFilterFNameOrLastName.Text = string.Empty;
-						dpFilterDOB.Text = string.Empty;
 					}
 				}
 
@@ -230,20 +240,25 @@ namespace Group2WPF
 		private void Delete_Click(object sender, RoutedEventArgs e)
 		{
 			try
-			{
-				if (ValidateData())
-				{
-					int id = Int32.Parse(txtID.Text);
-					Passenger? p = _passengerService.Get(id);
-					if (p == null)
+            {
+                var result = MessageBox.Show("Are you sure you want to Delete this flight?", "Confirm delete", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.Yes)
+                {
+					if (ValidateData())
 					{
-						MessageBox.Show("Not found passenger", "Fail!");
-					}
-					else
-					{
-						_passengerService.Delete(p);
-						txtID.IsReadOnly = false;
-						ResetInput();
+						int id = Int32.Parse(txtID.Text);
+						Passenger? p = _passengerService.Get(id);
+						if (p == null)
+						{
+							MessageBox.Show("Not found passenger", "Fail!");
+						}
+						else
+						{
+							_passengerService.Delete(p);
+							MessageBox.Show("Delete successfully", "Success");
+							txtID.IsReadOnly = false;
+							ResetInput();
+						}
 					}
 				}
